Implemeted a method of security to save the username in a specific object.
Implemented the Identity object and its Interface.

Registered is Interface to the object in the NinjectWebCommon file so that Ninject can resolve these dependencies.

Added a base controller class to house all the common code that we will have for every controller.

reated seperate controllers for measures, diaries, and foods.

added models for foods, measures and diaries.