Added the DiaryEntriesModel, DiaryEntrieController on my own following the pattern that the food and measures followed.
First add a new route in the webApiConfig that matches the route you need to take to access the entries.

Then make the controller, with 2 methods, 1 to get all the entries for one day and 1 to get a specific entry.

To support the DiaryEntriesController I have to make a DiaryEntryModel to hold all the data that is required.

Then added the create methods inside the modelFactory, when making this method you have to modify the existing Diary create aswell becuase you have to supply the diary wirth entries.

then supplying the diaryentries with data is just like creating the MeasureModel, it depends a lot on the diary and you get most of the data from the diaryEntry Entity.