Implemented the Post method in the API.
Now this one was a little complex, first you send the actual data you want to put into your server in the body of the httprequest.

Then you parse that into the object type that you want.

The parsing bascially matches the data that was send in the body with the data in the database and creates a new diaryentry object correspoing to the measure Id that is extracted from the http body.

from that measure Id, we get the measure from the repository, as well as the connected food item to that measure.

with those details we just parse the quantity and now we have all the information to build a new instance of DiaryEntry. Which we do and pass it to the controller.

the controller checks if the returned diary entry null then gets the diary of the current user.

check if the diary entry has any other entries for the same measure.

then if there are none, we add the entry to the diary.Entries property of the user's diary and call the saveall method.