Implemented a put/path method into the DieryEntriesController to update the values of the diary entries.
we use this method to update a portion of an object in a database. so in this case we are updating the value of quantity related to the object.

The other thing we did was to implement paging in the foodscontroller. We did it very manually by getting the max count size from the repository and defining a page size in our code.

We then used these values and the skip operator to show us the paged results.

next we improved the paging by providing the number of pages in total and links to the next page and the previous page.

These Urls were created using the UrlHelper object by passing the request object to it. All we did was told it that we need to use the food api route and gave it the page number and voila, it gave us page links! :O

Crazy... I need to know how that happened....

Finally we returned all that data to the client using an anonymous type.