In this part we made a RemoteProcedure style call to the webAPI.
we started with a DiarySummaryController which had a get method to return the summary of that day.

We then added the route neede to reach this controller in the webApiConfig file.

Finally we implemented the CreateSummary method on the ModelFactory.

The createsummary basicalle takes in the diary of an user and computes the total calorie consumption of the userfor that day.

I guess the use of this would be to let the server do the processing to find out how many calories have been eaten, rather than doing that one the client.

Seems fairly straightforward to me :)