Made changed to support requireing SSL to access the webapi.
Added a line in the webApiConfig file to add a filter i created to support SSL.

Created a filter called requireHttpsAttribute. which derived off of AuthorizationFilterAttribute class.

I used the OnAuthorized method to force the request to require https to connect.

The rest of the congiguration must be done on the server, IIS needs to be configured to use an SSL connection with SSL certificates and what not.