Explored token based authorization for developers who want to use an API that I am making.
Created the token model and added a post method to handle the creation of tokens for developers.

The method to create a new token using the secret and the appid is basic and for demonstraion only. But I get the idea.

This token, after it has been created should be able to be used to authorize a developer to log into and use the API.

TheModelFactory was used to crate a tokenModel so that we can only send the token and the expiration time to the use who requested the token in the first place.