This was alot more difficult that I had imagined,
Now that I had made the tokens for the developers, I had to create an authorizatioon process to validate the tokens that they were giving me.

This is a lot harder that it seemed.

First I modified the exisitng CountingKsAuthorizeAttribute by adding a per user bool to tell us if we want to authorize per user or not.

Then we refactored the code to first authorize based on the token and then later based on the user.

In order to authorize based on the token we extractedthe token and the appkey from the response query string.

then we looked into the repository to get the saved token in the databae. If this token id matches the appid and the token is not yet expired, the authorization is successful.

BUT!!!! In order to use the repository from here we need to use dependency injection. So we marked the repo with the inject attriubute, and then created the NinjectWebApiFilterProvider which used the ninject kernel to inject all the filters for actions and controllers that were needed.

then finally we had to add this new Filter to the globalconfiguration for ninject and pass in the filter object we made,. This was don in the NinjectWebCommon file..

All in all, quite exhauting. Dependency injection is still somewhat greek to me...