implemented versioning by creating a CountingKsControllerSelector which derives off of the DefaultControllerSelector class.
We overrided the select controller method to search for the controller name and the if successful to look for the veriosned controller name. if the latter was successful the method would return a httpcontrollerdescriptor which points to the versioned controller. If it was unsuccessful it would just point to the general controller.

This has to be added to the webapi configuration in webapiconfig.cs file. we need to replace the current HttpControllerSelector with our one.