Added a way to send the version requested via the query string of the url.
This feature was added into the controllerSelector's selectController method that we overrode.

the version information was sent through the query string with the key "v". So if there is a version selected it will pass that information on to the selectController method, if not it will default to 1.

to update the values shown in the URLs, the version number was hard coded into the urls in the model facory.