Implemented caching using the CacheCow projects's cache handler.
used the sql server to store the etags so that they are persistant etags and not weak etags.

connected the cachecow to out sql server using the entity framework.

Ran some sql scripts to generate some stored procedures and tables that cachecow needed to run.

Now using the etags and the header If-None-Match: "0c794867831d473787168277e3b6c200" we can check if the client object is uptodate or not.